Пытаюсь прикрутить Apache Ignite и Hibernate OGM

+++ application/api/src/main/java/ru/ulmc/bank/bean/IPrice.java
@@ -1,11 +1,12 @@
 package ru.ulmc.bank.bean;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public interface IPrice extends Serializable {
     int getVolume();
 
-    Double getBid();
+    BigDecimal getBid();
 
-    Double getOffer();
+    BigDecimal getOffer();
 }
